refactor: 상품주문 패키지를 주문(order) 패키지에 통합
- 애그리게이트가 다르다고 꼭 다른 패키지에 위치시키는 것이 아님. 둘은 애그리게이트가 다른 것이지 하위 도메인은 모두 주문으로 똑같음
