+++ src/main/java/kr/tatine/manibogo_oms_v2/common/model/Address.java
@@ -1,4 +1,4 @@
-package kr.tatine.manibogo_oms_v2.fulfillment.command.domain.order;
+package kr.tatine.manibogo_oms_v2.common.model;
 
 import jakarta.persistence.Embeddable;
 import lombok.AccessLevel;
