--- src/main/java/kr/tatine/manibogo_oms_v2/fulfillment/command/domain/item_order/AlreadyShippedException.java
@@ -1,3 +1,0 @@
-package kr.tatine.manibogo_oms_v2.fulfillment.command.domain.item_order;
-
-public class AlreadyShippedException extends RuntimeException { }
