--- src/main/java/kr/tatine/manibogo_oms_v2/fulfillment/command/domain/item_order/ItemOrderRepository.java
@@ -1,5 +1,0 @@
-package kr.tatine.manibogo_oms_v2.fulfillment.command.domain.item_order;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface ItemOrderRepository extends CrudRepository<ItemOrder, ItemOrderNumber> { }
