--- src/main/java/kr/tatine/manibogo_oms_v2/fulfillment/command/domain/order/OrderRepository.java
@@ -1,5 +1,0 @@
-package kr.tatine.manibogo_oms_v2.fulfillment.command.domain.order;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface OrderRepository extends CrudRepository<Order, OrderNumber> { }
