+++ src/main/java/kr/tatine/manibogo_oms_v2/fulfillment/command/domain/order/exception/AlreadyDispatchedException.java
@@ -1,0 +1,3 @@
+package kr.tatine.manibogo_oms_v2.fulfillment.command.domain.order.exception;
+
+public class AlreadyDispatchedException extends RuntimeException { }
