+++ src/main/java/kr/tatine/manibogo_oms_v2/fulfillment/command/domain/order/model/vo/ChargeType.java
@@ -1,4 +1,4 @@
-package kr.tatine.manibogo_oms_v2.fulfillment.command.domain.order;
+package kr.tatine.manibogo_oms_v2.fulfillment.command.domain.order.model.vo;
 
 import kr.tatine.manibogo_oms_v2.common.model.Describable;
 import lombok.Getter;
@@ -6,10 +6,11 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum SalesChannel implements Describable {
+public enum ChargeType implements Describable {
 
-    SMART_STORE("스마트스토어"),
-    LOCAL("매장");
+    COD("착불"),
+    PREPAID("선결제");
 
     private final String description;
+
 }
