+++ src/main/java/kr/tatine/manibogo_oms_v2/fulfillment/command/domain/order/model/vo/ShippingMethod.java
@@ -1,4 +1,4 @@
-package kr.tatine.manibogo_oms_v2.fulfillment.command.domain.item_order;
+package kr.tatine.manibogo_oms_v2.fulfillment.command.domain.order.model.vo;
 
 import kr.tatine.manibogo_oms_v2.common.model.Describable;
 import lombok.Getter;
@@ -6,11 +6,9 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum ChargeType implements Describable {
+public enum ShippingMethod implements Describable {
 
-    COD("착불"),
-    PREPAID("선결제");
+    DIRECT("직접전달"), PARCEL("택배,등기,소포");
 
     private final String description;
-
 }
